canal add extra field
